Temporarily remove CJS module support
It turns out that for CJS require to work with "type": "module",
.js files need to be all renamed to .cjs,
as well as all references to them,
meaning strings passed to require also need modifications.
While doable, I don't want to spend that much amount of effort just to
support CJS modules.

Maybe we can use rollup for CJS to minimize the workload?